+++ lib/grax/schema/property.ex
@@ -1,7 +1,9 @@
 defmodule Grax.Schema.Property do
   @moduledoc false
 
-  @shared_attrs [:schema, :name, :iri, :type]
+  alias Grax.Schema.Type
+
+  @shared_attrs [:schema, :name, :iri, :type, :cardinality]
 
   def shared_attrs, do: @shared_attrs
 
@@ -17,16 +19,27 @@ defmodule Grax.Schema.Property do
   defp normalize_iri(iri), do: RDF.iri!(iri)
 
   def value_set?(%{type: type}), do: value_set?(type)
-  def value_set?({:list_set, _}), do: true
-  def value_set?(_), do: false
+  def value_set?(type), do: Type.set?(type)
 
   def default({:list_set, _}), do: []
   def default(_), do: nil
 
-  def init_type(name, type, property_type) do
-    case internal_type(type, property_type) do
+  def type_with_cardinality(name, opts, property_type) do
+    type_with_cardinality(
+      name,
+      opts[:type],
+      Keyword.get(opts, :required, false),
+      property_type
+    )
+  end
+
+  def type_with_cardinality(name, type, false, property_type) do
+    case value_type(type, property_type) do
+      {:ok, type, card} ->
+        {type, card}
+
       {:ok, type} ->
-        type
+        {type, nil}
 
       {:error, nil} ->
         raise ArgumentError, "invalid type definition #{inspect(type)} for property #{name}"
@@ -36,14 +49,30 @@ defmodule Grax.Schema.Property do
     end
   end
 
-  def internal_type({:list_set, type}, property_type) do
-    with {:ok, inner_type} <- property_type.internal_type(type) do
-      {:ok, {:list_set, inner_type}}
+  def type_with_cardinality(name, type, true, property_type) do
+    with {type, cardinality} <- type_with_cardinality(name, type, false, property_type) do
+      cond do
+        not Type.set?(type) ->
+          {type, 1}
+
+        is_nil(cardinality) ->
+          {type, {:min, 1}}
+
+        true ->
+          raise ArgumentError,
+                "property #{name}: required option is not allowed when cardinality constraints are given"
+      end
     end
   end
 
-  def internal_type(type, property_type) do
-    property_type.internal_type(type)
+  defp value_type({:list_set, type, cardinality}, property_type) do
+    with {:ok, inner_type} <- property_type.value_type(type) do
+      {:ok, {:list_set, inner_type}, cardinality}
+    end
+  end
+
+  defp value_type(type, property_type) do
+    property_type.value_type(type)
   end
 end
 
@@ -54,26 +83,26 @@ defmodule Grax.Schema.DataProperty do
   alias Grax.Datatype
   alias RDF.Literal
 
-  defstruct Property.shared_attrs() ++ [:required, :default, :from_rdf, :to_rdf]
+  defstruct Property.shared_attrs() ++ [:default, :from_rdf, :to_rdf]
 
   @default_type :any
 
   def new(schema, name, iri, opts) do
-    type = Property.init_type(name, opts[:type], __MODULE__)
+    {type, cardinality} = Property.type_with_cardinality(name, opts, __MODULE__)
 
     __MODULE__
     |> Property.init(schema, name, iri, opts)
     |> struct!(
       type: type,
+      cardinality: cardinality,
       default: init_default(type, opts[:default]),
-      required: Keyword.get(opts, :required, false),
       from_rdf: normalize_custom_mapping_fun(opts[:from_rdf], schema),
       to_rdf: normalize_custom_mapping_fun(opts[:to_rdf], schema)
     )
   end
 
-  def internal_type(nil), do: internal_type(@default_type)
-  def internal_type(type), do: Datatype.get(type)
+  def value_type(nil), do: value_type(@default_type)
+  def value_type(type), do: Datatype.get(type)
 
   defp init_default(type, nil), do: Property.default(type)
 
@@ -105,18 +134,18 @@ defmodule Grax.Schema.LinkProperty do
   defstruct Property.shared_attrs() ++ [:preload, :on_type_mismatch]
 
   def new(schema, name, iri, opts) do
+    {type, cardinality} = Property.type_with_cardinality(name, opts, __MODULE__)
+
     cond do
       Keyword.has_key?(opts, :default) ->
         raise ArgumentError, "the :default option is not supported on links"
-
-      Keyword.has_key?(opts, :required) ->
-        raise ArgumentError, "the :required option is not supported on links"
 
       true ->
         __MODULE__
         |> Property.init(schema, name, iri, opts)
         |> struct!(
-          type: Property.init_type(name, opts[:type], __MODULE__),
+          type: type,
+          cardinality: cardinality,
           preload: opts[:preload],
           on_type_mismatch: init_on_type_mismatch(opts[:on_type_mismatch])
         )
@@ -135,9 +164,9 @@ defmodule Grax.Schema.LinkProperty do
           })"
   end
 
-  def internal_type(nil), do: {:error, "type missing"}
+  def value_type(nil), do: {:error, "type missing"}
 
-  def internal_type(class_mapping) when is_map(class_mapping) do
+  def value_type(class_mapping) when is_map(class_mapping) do
     {:ok,
      {:resource,
       Map.new(class_mapping, fn
@@ -146,7 +175,7 @@ defmodule Grax.Schema.LinkProperty do
       end)}}
   end
 
-  def internal_type(schema), do: {:ok, {:resource, schema}}
+  def value_type(schema), do: {:ok, {:resource, schema}}
 
   def default(%__MODULE__{} = link_schema) do
     %Link.NotLoaded{