+++ CHANGELOG.md
@@ -12,10 +12,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - heterogeneous link properties which can link different types of resources
   to different schemas
 - schema inheritance
+- support for cardinality constraints on properties
+- support for `:required` on link properties  
 - support for custom `:from_rdf` mappings on custom fields  
-- support for custom mappings on separate modules
-- the `build` functions can now be called with a single map when it contains an id
-  in an `:__id__` field
+- support for custom mappings in separate modules
+- the `build` functions can now be called with a single map when the map contains
+  an id an `:__id__` field
 
 ### Changed
 
@@ -25,6 +27,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - the default value of link properties has been changed to `nil` respective the empty list
   (previously it was a `Grax.Link.NotLoaded` struct, which is now set explicitly 
   during loading)
+- failing `:required` requirements result in a `Grax.Schema.CardinalityError` instead
+  of a `Grax.Schema.RequiredPropertyMissing` exception
 
 
 [Compare v0.1.0...HEAD](https://github.com/rdf-elixir/grax/compare/v0.1.0...HEAD)